Divided UserDataModels into separate Local Online, and shared interfaces
